+++ lib/sorcery/controller/submodules/http_basic_auth.rb
@@ -45,9 +45,9 @@ module Sorcery
           # given to main controller module as a login source callback
           def login_from_basic_auth
             authenticate_with_http_basic do |username, password|
-              @current_user = (user_class.authenticate(username, password) if session[:http_authentication_used]) || false
-              auto_login(@current_user) if @current_user
-              @current_user
+              @current_sorcerer = (user_class.authenticate(username, password) if session[:http_authentication_used]) || false
+              auto_login(@current_sorcerer) if @current_sorcerer
+              @current_sorcerer
             end
           end